Updating FontAwesome Icons in Admin area
Some FontAwesome icons do not work properly when other extensions are added to Joomla. These changes are for compatability with newer versions of Joomla and FontAwesome 5.

Some important icons get broken sometimes
arrows for the spacer particle
TM (trademark) for the branding particle
For other particles that utilize the pencil, copy, trash FA icons

For the Branding particle located in:
engines\common\nucleus\particles
branding.yaml
branding.html.twig

In the Gantry 5 it has the class "fa fa-fw fa-trademark" on the left side menu of particles but "fas fa-fw fa-trademark" is more compatible. When the branding particle is added to a layout, it again has the class "fa fa-trademark", and changing it to "fas fa-trademark" fixes the problem. However, I'm unable to figure out where to make these changes for this particle so it displays properly.

I am uanble to pinpoint exactly what extensions trigger this issue but I do know making these changes fixes it and the changes are still compatible with a fresh install of Gantry 5. Just need to fix the Branding particle.